Revert "nxdlformat.xsl no longer needed"
This reverts commit 5e247fcf8cb29dc556a93fc9d0ab52e5e50768a7.

Rationale [Pete]:

The tree view of the NXDL files has proven to be the most useful description on which to base a discussion of the structure of a particular NeXus class definition. Inclusion of nxdlformat.xsl in each of the directories (applications, base_classes, contributed_definitions) enables the easy generation of a tree view of any .nxdl.xml file by dragging that file into an internet browser
(or double-clicking if .xml files are default opened by the internet browser).

This infrastructure is enabled by the second line in each NXDL file:
<?xml-stylesheet type="text/xsl" href="nxdlformat.xsl" ?>

The proposed commit breaks this infrastructure.