Don't render XML comments.
In response to Pete:

When rendering NXDL specifications, XML comments should not be rendered.

Such as, in NXmirror, do not render this comment (even though it is a worthy thing to be repaired in the NXDL file):

change:
(geometry): NXgeometry
<!– TODO explain what this group means –>

to:
(geometry): NXgeometry

--

Note however that there are also <doc> entries like
"TODO: documentation needed". Those are still rendered.