Removed trailing periods.
In response to Pete:

When the NXDL are rendered, the metadata often end with a period. These are not complete sentences and the period is not appropriate. In some cases, it makes the information correct. Such as:

Groups cited: NXdata, NXentry, NXinstrument, NXparameters, NXprocess, NXsample, NXsource.
History: Introduced in NeXus version 1.0b.
Source: Automatically generated from https://github.com/nexusformat/definitions/blob/master/applications/NXsqom.nxdl.xml.
2.4.15 NXtas
Category: application.
Description:
This is an application definition for a triple axis spectrometer. It is for the trademark scan of the TAS, the
Q-E scan. For your alignment scans use the rules in NXscan.
Extends: NXobject.

Of these, the trailing period should be removed from Groups cited, Source, Category, Extends