Revert "History" to "Version"; compact "Status" field as first metadata.
E.g.:

Status:

    base class, extends NXobject, version 1.0

In reponse to Pete (but avoiding a table, mainly for typographic parsimony):

For each NXDL, the version (/definition/@version attribute) is described incorrectly as History.

change:
History: Introduced in NeXus version 1.0b.

to:
Version: 1.0b

and place this as one of the first metadata about the NXDL.

It is clearer, more systematic, and more compact to render these first metadata items in a table:

====================== ======== =======
Category Extends Version
====================== ======== =======
application definition NXobject 1.0b
====================== ======== =======