No longer distinguish between NX_CHAR and (NX_CHAR).
In response to Pete:

Since NX_CHAR is the default type when not specified for a NXDL field, it is not necessary to use the word "untyped" when rendering fields of this type. The parentheses around the data type indicate that a default choice has been provided.

For example, in NXarchive, change:

definition: untyped (NX_CHAR)

to

definition: (NX_CHAR)

--

Since there is a clear rule that NX_CHAR is the default type,
there is no good reason to render the difference between
type=default and type=NC_CHAR at all.
Its only information content is about the laziness of person
who edited the NXDL page.