Build fails due to problems with nxdl2rst.py
The script determines the type of a class from the name of the subdirectory
where the XML files are stored. For the new build this is no longer working as
the files are generated in exactly this directory.
Need to change nxdl2rst.

Build of the schemafiles needs a total rewrite too. The current approach is far
too complex.