introduce NXsample_component
refs #433

This adds a NXsample_component parallel to the existing n_comp
n_comp is not yet deprecated, but NXsample_component is better.
With that you can record the total mass of the sample in addtion to
the one for the individual components. And you can specify values
that are known for one component only

This also fixes the ambigity between space and point group.
That make the change not fully "backward compatible".
But the old version was essentially broken.

Also added are unit_cell_a,b,c,alpha,beta,gamma because the
exiting unit_cell makes no sense with the length units.
Old style not deprecated yet.

Whether the unit_cell_* is useful for the NXsample_components does
require discussion