Replace 'dataset' with 'field'
Data sets can mean a multitude of things in different contexts. For example, in many data catalogs, data sets refer to collections of data objects. Even within this file, some might construe 'dataset' to refer to the plottable data. However, there are many uses of the word 'field', which is less ambiguous, so I would propose that this is used wherever we are referring to an item in the group.