Rename VARIABLE and AXISNAME
The use of 'VARIABLE' was presumably inspired by the term 'independent variable', but this is scarcely used in favor of 'dimension scale'. Since we are always referring to an axis, I suggest changing 'VARIABLE' to 'AXIS'.  Since both refer to an arbitrary name, the use of 'AXISNAME' is redundant. I also made a couple of references upper case for consistency.