Removed redundant text
I object to the phrase "file readers are encouraged to make their best efforts to plot the data" since it implies that the axis definitions become ambiguous. Since I have implemented a generic plotter that has never had a problem in identifying axes unambiguously, I know this is not true. I would suggest that the whole text is unnecessary, and will just confuse any outsiders not familiar with the controversy. All that is needed is a statement that the attributes should be provided, assuming that is still the NIAC recommendation.