Clarify note on IUCr coordinate system.
The comment on Z is correct, but in ImageCIF, the X axis is defined based on the goniometer orientation. It is common for goniometers to be mounted in different orientations, which will flip X and Y depending on the beamline. Therefore, the second part of the note has been removed.

Fixes #703