Better rewrite of NXdata scaling_factor and offset fields (#1343)
Adds FIELDNAME_scaling_factor and FIELDNAME_offset as fields to NXdata. These fields should be used instead of scaling_factor and offset, as it is ambiguous which fields to apply them to in the case of multiple signals. _scaling_factor and _offset are added as reserved suffixes.

Additionally adds a formula for how to apply these fields:

corrected values = (FIELDNAME + offset) * scaling_factor

For NXmx, adds clarification these fields can be used as background and gain correction fields, and defines these terms. Also the possible rank options are elaborated.

Closes #1332.

Also:
- Change NXdata scaling_factor to refer to "plotted" data
- Change NXdata to refer to "corrected" data, in addition to "physical" data, since it describes units of photons
- Use FIELDNAME_scaling_factor and FIELDNAME_offset.  This resolves ambiguity if there is more than one signal
- For NXmx specify data_scaling_factor and data_offset since the field data is named in the NXdata group
- NXmx: fully define pedestal, bias, and gain and how they relate to offset and scaling_factor
- Add reserved suffixes _scaling_factor and _offset
- NXdata: define better how the stored values in FIELDNAME are converted to physical values
- NXdata: more clarification on how scaling_factor and offset without FIELDNAME are ambiguous
- Add implied defaults for sclaing_factor and offset